requirements.txt
django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
pillow==10.2.0
python-dotenv==1.0.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
coverage==7.4.3
django-cors-headers==4.3.1
requests==2.31.0
stripe==8.7.0